+++ com.unity.render-pipelines.high-definition/Editor/Material/Fabric/ShaderGraph/FabricSubTarget.cs
@@ -10,6 +10,7 @@ using UnityEditor.ShaderGraph.Legacy;
 using UnityEditor.Rendering.HighDefinition.ShaderGraph.Legacy;
 using static UnityEngine.Rendering.HighDefinition.HDMaterialProperties;
 using static UnityEditor.Rendering.HighDefinition.HDShaderUtils;
+using static UnityEditor.Rendering.HighDefinition.HDFields;
 
 namespace UnityEditor.Rendering.HighDefinition.ShaderGraph
 {
@@ -22,7 +23,8 @@ namespace UnityEditor.Rendering.HighDefinition.ShaderGraph
         protected override string customInspector => "Rendering.HighDefinition.FabricGUI";
         protected override ShaderID shaderID => HDShaderUtils.ShaderID.SG_Fabric;
         protected override string subShaderInclude => "Packages/com.unity.render-pipelines.high-definition/Runtime/Material/Fabric/Fabric.hlsl";
-        protected override FieldDescriptor subShaderField => HDFields.SubShader.Fabric;
+        protected override string raytracingInclude => CoreIncludes.kFabricRaytracing;
+        protected override FieldDescriptor subShaderField => new FieldDescriptor(kSubShader, "Fabric SubShader", "");
         protected override bool requireSplitLighting => fabricData.subsurfaceScattering;
 
         FabricData m_FabricData;
@@ -39,6 +41,9 @@ namespace UnityEditor.Rendering.HighDefinition.ShaderGraph
             set => m_FabricData = value;
         }
 
+        public static FieldDescriptor CottonWool =              new FieldDescriptor(kMaterial, "CottonWool", "_MATERIAL_FEATURE_COTTON_WOOL 1");
+        public static FieldDescriptor Silk =                    new FieldDescriptor(kMaterial, "Silk", "_MATERIAL_FEATURE_SILK 1");
+
         protected override SubShaderDescriptor GetRaytracingSubShaderDescriptor()
         {
             var descriptor = base.GetRaytracingSubShaderDescriptor();
@@ -54,12 +59,11 @@ namespace UnityEditor.Rendering.HighDefinition.ShaderGraph
             base.GetFields(ref context);
 
             // Fabric specific properties
-            context.AddField(HDFields.CottonWool,                           fabricData.materialType == FabricData.MaterialType.CottonWool);
-            context.AddField(HDFields.Silk,                                 fabricData.materialType == FabricData.MaterialType.Silk);
-            context.AddField(HDFields.SubsurfaceScattering,                 fabricData.subsurfaceScattering && systemData.surfaceType != SurfaceType.Transparent);
-            context.AddField(HDFields.Transmission,                         fabricData.transmission);
-            context.AddField(HDFields.DoAlphaTest,                          systemData.alphaTest && context.pass.validPixelBlocks.Contains(BlockFields.SurfaceDescription.AlphaClipThreshold));
-            context.AddField(HDFields.EnergyConservingSpecular,             fabricData.energyConservingSpecular);
+            context.AddField(CottonWool,                           fabricData.materialType == FabricData.MaterialType.CottonWool);
+            context.AddField(Silk,                                 fabricData.materialType == FabricData.MaterialType.Silk);
+            context.AddField(SubsurfaceScattering,                 fabricData.subsurfaceScattering && systemData.surfaceType != SurfaceType.Transparent);
+            context.AddField(Transmission,                         fabricData.transmission);
+            context.AddField(EnergyConservingSpecular,             fabricData.energyConservingSpecular);
         }
 
         public override void GetActiveBlocks(ref TargetActiveBlockContext context)